requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.4.0
pymongo==3.12.1
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
